python-backend/requirements.txt
whisperx==3.8.2
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
torch==2.8.0
python-dotenv>=1.0.0
